✨ 클래스 필터링, 정렬 기능 구현

+++ src/main/java/com/tms/sportlight/dto/CourseCardDTO.java
@@ -23,6 +23,6 @@ public class CourseCardDTO {
   private CourseLevel level;
   private String category;
   private double rating;
-  private int reviewCount;
+  private Long reviewCount;
 
 }
